+++ src/context-menu.js
@@ -1,31 +1,38 @@
-import {Config} from './config';
-import css from 'context-menu.scss';
+import { Config } from './config';
+import css from './context-menu.scss';
 
 export default class ContextMenu {
-
-  get left() { return this._containerDiv.getBoundingClientRect().left; }
-  get top() { return this._containerDiv.getBoundingClientRect().top; }
-  get width() { return this._containerDiv.getBoundingClientRect().width; }
-  get height() { return this._containerDiv.getBoundingClientRect().height; }
+  get left() {
+    return this._containerDiv.getBoundingClientRect().left;
+  }
+  get top() {
+    return this._containerDiv.getBoundingClientRect().top;
+  }
+  get width() {
+    return this._containerDiv.getBoundingClientRect().width;
+  }
+  get height() {
+    return this._containerDiv.getBoundingClientRect().height;
+  }
 
   constructor(div, options) {
     this._defaultOptions = {
-      width: 150
+      width: 150,
     };
     this._options = Object.assign({}, this._defaultOptions, options || {});
     this._parentDiv = div;
-    this._containerDiv = document.createElement("div");
+    this._containerDiv = document.createElement('div');
     this._containerDiv.classList.add(`${Config.prefix}-context-menu`);
     this._containerDiv.stylesheet = css;
-    this._containerDiv.oncontextmenu = event => event.preventDefault();
-    this._list = document.createElement("ul");
+    this._containerDiv.oncontextmenu = (event) => event.preventDefault();
+    this._list = document.createElement('ul');
     this._containerDiv.appendChild(this._list);
     div.appendChild(this._containerDiv);
     this._isVisible = false;
   }
 
   addItem(name, isVisible, cb, context = this) {
-    let item = document.createElement("li");
+    let item = document.createElement('li');
     let content = document.createTextNode(name);
 
     item.appendChild(content);
@@ -52,8 +59,8 @@ export default class ContextMenu {
   }
 
   toggleItems(itemsToShow, itemsToHide) {
-    itemsToShow.forEach(item => this.showItem(item));
-    itemsToHide.forEach(item => this.hideItem(item));
+    itemsToShow.forEach((item) => this.showItem(item));
+    itemsToHide.forEach((item) => this.hideItem(item));
   }
 
   show(point) {
@@ -64,12 +71,14 @@ export default class ContextMenu {
       position: absolute;
       width: ${this._options.width}px; 
     `;
-    let isXOverflow = this._parentDiv.getBoundingClientRect().width <= point.x + this.width;
-    let isYOverflow = this._parentDiv.getBoundingClientRect().height <= point.y + this.height;
+    let isXOverflow =
+      this._parentDiv.getBoundingClientRect().width <= point.x + this.width;
+    let isYOverflow =
+      this._parentDiv.getBoundingClientRect().height <= point.y + this.height;
 
     this._containerDiv.style.cssText += `
-      ${isXOverflow ? "right: 0px;" : "left: " + point.x + "px;"}
-      ${isYOverflow ? "bottom: 14px;" : "top: " + point.y + "px;"}
+      ${isXOverflow ? 'right: 0px;' : 'left: ' + point.x + 'px;'}
+      ${isYOverflow ? 'bottom: 14px;' : 'top: ' + point.y + 'px;'}
       visibility: visible;
     `;
   }
@@ -86,4 +95,4 @@ export default class ContextMenu {
       this.show(point);
     }
   }
-};
+}