+++ src/index.js
@@ -1,28 +1,49 @@
-import css from 'index.scss';
+import {drag} from 'd3-drag';
+import {select, selectAll} from 'd3-selection';
 import {Config} from './config';
 import ContextMenu from './context-menu';
 import Tooltip from './tooltip';
-import {select, selectAll, event} from 'd3-selection';
 import ProjectionUtility from './projection-utility';
-import {Geometry} from 'geometry';
-import {Segment} from 'segment';
-import {drag} from 'd3-drag';
+import {Geometry} from './geometry';
+import {Segment} from './segment';
 import Helper from './helper';
 import {UnitTypeId} from './UnitTypeId';
 import {EVENT_START, EVENT_END, EVENT_CHANGE} from './events';
 import {ObjectAssign} from './polyfills';
-import {deepClone} from "./utils";
+import {deepClone, getClass} from './utils';
+import './index.scss';
 
 export default class MeasureTool {
+  get lengthText() {
+    return this._helper.formatLength(this._length || 0);
+  }
 
-  get lengthText() { return this._helper.formatLength(this._length || 0); };
-  get areaText() { return this._helper.formatArea(this._area || 0); };
-  get length() { return this._length || 0; };
-  get area() { return this._area || 0; };
-  get segments() { return deepClone(this._segments) || []; };
-  get points() { return deepClone(this._geometry.nodes.map(x=> ({lat: x[1], lng: x[0]}))) || []; };
+  get areaText() {
+    return this._helper.formatArea(this._area || 0);
+  }
 
-  static get UnitTypeId() { return UnitTypeId};
+  get length() {
+    return this._length || 0;
+  }
+
+  get area() {
+    return this._area || 0;
+  }
+
+  get segments() {
+    return deepClone(this._segments) || [];
+  }
+
+  get points() {
+    return (
+      deepClone(this._geometry.nodes.map((x) => ({lat: x[1], lng: x[0]}))) ||
+      []
+    );
+  }
+
+  static get UnitTypeId() {
+    return UnitTypeId;
+  }
 
   /**
    * Creates a new measure tool for the google.maps.Map instance.
@@ -38,13 +59,16 @@ export default class MeasureTool {
       contextMenu: true,
       tooltip: true,
       unit: UnitTypeId.METRIC,
-      initialSegments: []
+      initialSegments: [],
+      language: navigator ? navigator.language : 'en',
+      invertColor: false,
+      ...options,
     };
-    Object.assign(this._options, options);
     this._map = map;
     this._map.setClickableIcons(false);
     this._id = Helper.makeId(4);
     this._events = new Map();
+    this._geometry = new Geometry();
     this._init();
   }
 
@@ -53,12 +77,22 @@ export default class MeasureTool {
   }
 
   _init() {
-    this._containerDiv = this._map.getDiv().querySelector("div:first-child");
+    this._containerDiv = this._map.getDiv().querySelector('div:first-child');
 
     if (this._options.contextMenu) {
-      this._contextMenu = new ContextMenu(this._containerDiv, { width: 160 });
-      this._startElementNode = this._contextMenu.addItem("Measure distance", true, this.start, this);
-      this._endElementNode = this._contextMenu.addItem("Clear measurement", false, this.end, this);
+      this._contextMenu = new ContextMenu(this._containerDiv, {width: 160});
+      this._startElementNode = this._contextMenu.addItem(
+        'Measure distance',
+        true,
+        this.start,
+        this
+      );
+      this._endElementNode = this._contextMenu.addItem(
+        'Clear measurement',
+        false,
+        this.end,
+        this
+      );
       this._bindToggleContextMenu();
     }
 
@@ -67,26 +101,42 @@ export default class MeasureTool {
     }
 
     this._helper = new Helper({
-      unit: this._options.unit
+      unit: this._options.unit,
     });
     this._initOverlay();
   }
 
   _bindToggleContextMenu() {
-    this._map.addListener('rightclick', mouseEvent => {
-      this._firstClick = mouseEvent;
-      this._contextMenu.show(this._projection.fromLatLngToContainerPixel(mouseEvent.latLng));
+    this._containerDiv.addEventListener('touchstart', (touchEvent) => {
+      console.debug('touched', touchEvent);
+      if (touchEvent.targetTouches.length === 2) {
+        console.debug('2 finger touches');
+      }
     });
-    document.addEventListener('keydown', event => {
-      if (event.which === 27) {
+    this._containerDiv.addEventListener('touchend', (touchEvent) => {
+      console.debug('touch end', touchEvent);
+      if (touchEvent.targetTouches.length === 2) {
+        console.debug('2 finger touches');
+      }
+    });
+    this._map.addListener('contextmenu', (mouseEvent) => {
+      this._firstClick = mouseEvent;
+      this._contextMenu.show(
+        this._projection.fromLatLngToContainerPixel(mouseEvent.latLng)
+      );
+    });
+    document.addEventListener('keydown', (event) => {
+      if (event.key === 'Escape' || event.which === 27) {
         this._contextMenu.hide();
       }
     });
-    this._containerDiv.addEventListener('mousedown', event => {
-      if (event.clientX >= this._contextMenu.left &&
-          event.clientX <= this._contextMenu.left + this._contextMenu.width &&
-          event.clientY >= this._contextMenu.top &&
-          event.clientY <= this._contextMenu.top + this._contextMenu.height) {
+    this._containerDiv.addEventListener('mousedown', (event) => {
+      if (
+        event.clientX >= this._contextMenu.left &&
+        event.clientX <= this._contextMenu.left + this._contextMenu.width &&
+        event.clientY >= this._contextMenu.top &&
+        event.clientY <= this._contextMenu.top + this._contextMenu.height
+      ) {
         return;
       }
       this._contextMenu.hide();
@@ -117,15 +167,19 @@ export default class MeasureTool {
 
     if (this._options.contextMenu && this._firstClick) {
       this._checkClick(this._firstClick);
-      this._contextMenu.toggleItems([this._endElementNode], [this._startElementNode]);
+      this._contextMenu.toggleItems(
+        [this._endElementNode],
+        [this._startElementNode]
+      );
     }
 
-    this._mapClickEvent = this._map.addListener('click', mouseEvent => this._checkClick(mouseEvent));
-    // this._mapZoomChangedEvent = this._map.addListener('zoom_changed', () => this._redrawOverlay());
+    this._mapClickEvent = this._map.addListener('click', (mouseEvent) =>
+      this._checkClick(mouseEvent)
+    );
     this._map.setOptions({draggableCursor: 'default'});
     this._started = true;
 
-    if (typeof this._events.get(EVENT_START) === "function") {
+    if (typeof this._events.get(EVENT_START) === 'function') {
       this._events.get(EVENT_START)();
     }
   }
@@ -136,16 +190,17 @@ export default class MeasureTool {
   end() {
     if (!this._started) return;
 
-    if (typeof this._events.get(EVENT_END) === "function") {
+    if (typeof this._events.get(EVENT_END) === 'function') {
       this._events.get(EVENT_END)(this._getResults());
     }
 
     if (this._options.contextMenu) {
-      this._contextMenu.toggleItems([this._startElementNode], [this._endElementNode]);
-
+      this._contextMenu.toggleItems(
+        [this._startElementNode],
+        [this._endElementNode]
+      );
     }
     this._mapClickEvent.remove();
-    // this._mapZoomChangedEvent.remove();
     this._geometry = new Geometry();
     this._onRemoveOverlay();
     this._setOverlay();
@@ -178,7 +233,7 @@ export default class MeasureTool {
    * @param value - value to set
    */
   setOption(option, value) {
-    if (!this._options[option]) {
+    if (this._options[option] === undefined) {
       throw new Error(`${option} is not a valid option on MeasureTool`);
     }
 
@@ -195,7 +250,6 @@ export default class MeasureTool {
       // only do this if there is actually an overlay to re-render
       this._redrawOverlay();
     }
-
   }
 
   _initOverlay() {
@@ -217,53 +271,58 @@ export default class MeasureTool {
     }
 
     this._projection = this._overlay.getProjection();
-    this._projectionUtility = new ProjectionUtility(this._containerDiv, this._projection);
+    this._projectionUtility = new ProjectionUtility(
+      this._containerDiv,
+      this._projection
+    );
     // Add svg to Pane
     this._svgOverlay = select(this._overlay.getPanes().overlayMouseTarget)
       .append('div')
-        .attr('class',`${Config.prefix}-measure-points-${this._id}`)
+      .attr('class', `${Config.prefix}-measure-points-${this._id}`)
       .append('svg')
-        .attr('class',`${Config.prefix}-svg-overlay`);
+      .attr('class', `${Config.prefix}-svg-overlay`);
 
-    this._linesBase = this._svgOverlay
-      .append('g').attr('class', 'base');
-    this._linesBase
-      .selectAll("line")
-      .data(this._geometry ? this._geometry.lines : []);
+    this._svgOverlay.append('defs').append('filter')
+      .attr('id', 'circle-shadow')
+      .append('feDropShadow')
+      .attr('dx', 0)
+      .attr('dy', 0)
+      .attr('stdDeviation', 0.5)
+      .attr('flood-color', 'white');
 
-    this._linesAux = this._svgOverlay
-      .append('g').attr('class', 'aux');
-    this._linesAux
-      .selectAll("line")
-      .data(this._geometry ? this._geometry.lines : []);
+    this._linesBase = this._svgOverlay.append('g').attr('class', 'base');
+    this._linesBase.selectAll('line').data(this._geometry.lines);
 
-    this._nodeCircles =  this._svgOverlay
-      .append('g').attr('class', 'node-circle');
-    this._nodeCircles
-      .selectAll('circle')
-      .data(this._geometry ? this._geometry.nodes : []);
+    this._linesAux = this._svgOverlay.append('g').attr('class', 'aux');
+    this._linesAux.selectAll('line').data(this._geometry.lines);
+
+    this._lineDrag = this._svgOverlay.append('g').attr('class', 'drag');
+    this._lineDrag.selectAll('line').data(this._geometry.lines);
+
+    this._nodeCircles = this._svgOverlay
+      .append('g')
+      .attr('class', 'node-circle');
+    this._nodeCircles.selectAll('circle').data(this._geometry.nodes);
 
     if (this._options.showSegmentLength) {
       this._segmentText = this._svgOverlay
-        .append('g').attr('class', 'segment-text');
-      this._segmentText
-        .selectAll('text')
-        .data(this._geometry ? this._geometry.lines : []);
+        .append('g')
+        .attr('class', 'segment-text');
+      this._segmentText.selectAll('text').data(this._geometry.lines);
     }
 
     if (this._options.showAccumulativeLength) {
-      this._nodeText = this._svgOverlay
-        .append('g').attr('class', 'node-text');
-      this._nodeText
-        .selectAll('text')
-        .data(this._geometry ? this._geometry.nodes : []);
+      this._nodeText = this._svgOverlay.append('g').attr('class', 'node-text');
+      this._nodeText.selectAll('text').data(this._geometry.nodes);
     }
 
     this._hoverCircle = this._svgOverlay
-      .append('g').attr('class', 'hover-circle');
+      .append('g')
+      .attr('class', 'hover-circle');
     this._hoverCircle
-      .append("circle")
-      .attr('class', 'grey-circle')
+      .append('circle')
+      .style('filter', 'url(#circle-shadow)')
+      .attr('class', getClass('grey-circle', this._options.invertColor))
       .attr('r', 5);
 
     if (this._initComplete && !this._started) {
@@ -285,8 +344,10 @@ export default class MeasureTool {
       this._updateNodeText();
     }
     if (this._geometry) {
-      this._updateArea(this._geometry.nodes.length - 1,
-        this._geometry.nodes[this._geometry.nodes.length - 1]);
+      this._updateArea(
+        this._geometry.nodes.length - 1,
+        this._geometry.nodes[this._geometry.nodes.length - 1]
+      );
     }
     this._dispatchMeasureEvent();
   }
@@ -306,10 +367,13 @@ export default class MeasureTool {
     this._overlay.draw();
   }
 
-  _checkClick(mouseEvent){
+  _checkClick(mouseEvent) {
     // Use circle radius 'r' as a flag to determine if it is a delete or add event.
-    if(!this._dragged && this._nodeCircles.selectAll('circle[r="6"]').size() == 0 &&
-       !this._hoverCircle.select("circle").attr('cx')) {
+    if (
+      !this._dragged &&
+      this._nodeCircles.selectAll('circle[r="6"]').size() === 0 &&
+      !this._hoverCircle.select('circle').attr('cx')
+    ) {
       const node = [mouseEvent.latLng.lng(), mouseEvent.latLng.lat()];
       this._geometry.addNode(node);
       this._overlay.draw();
@@ -320,154 +384,200 @@ export default class MeasureTool {
   _updateCircles() {
     let self = this;
     // join with old data
-    let circles = this._nodeCircles.selectAll("circle")
-      .data(this._geometry ? this._geometry.nodes : [])
-        .attr('class', (d, i) => i === 0 ? 'cover-circle head-circle' : 'cover-circle')
-        .attr('r', 5)
-        .attr('cx', d => this._projectionUtility.latLngToSvgPoint(d)[0])
-        .attr('cy', d => this._projectionUtility.latLngToSvgPoint(d)[1])
-        .on('mouseover', function(d, i){ self._onOverCircle(d, i, this);})
-        .on('mouseout', function(d){ self._onOutCircle(d, this);})
-        .on('touchstart', function(d, i){ self._onOverCircle(d, i, this);})
-        .on('touchleave', function(d){ self._onOutCircle(d, this);})
-        .on('mousedown', () => this._hideTooltip())
-        .call(this._onDragCircle());
+    let circles = this._nodeCircles
+      .selectAll('circle')
+      .data(this._geometry.nodes)
+      .join('circle')
+      .datum((d, i) => [d, i])
+      .style('filter', 'url(#circle-shadow)')
+      .attr('class', ([, i]) =>
+        i === 0 ? `${getClass('cover-circle', this._options.invertColor)} head-circle` : getClass('cover-circle', this._options.invertColor)
+      )
+      .attr('r', 5)
+      .attr('cx', ([d]) => this._projectionUtility.latLngToSvgPoint(d)[0])
+      .attr('cy', ([d]) => this._projectionUtility.latLngToSvgPoint(d)[1])
+      .on('mouseover', function (event, [d, i]) {
+        self._onOverCircle(d, i, this);
+      })
+      .on('mouseout', function (event, [d]) {
+        self._onOutCircle(d, this);
+      })
+      .on('touchstart', function (event, [d, i]) {
+        event.preventDefault();
+        self._onOverCircle(d, i, this, true);
+      })
+      .on('touchend', function (event, [d]) {
+        event.preventDefault();
+        self._onOutCircle(d, this);
+      })
+      .on('mousedown', () => this._hideTooltip())
+      .call(this._onDragCircle());
 
     // enter and seat the new data with same style.
     circles
       .enter()
       .append('circle')
-        .attr('class', 'cover-circle')
-        .attr('r', 5)
-        .attr('cx', d => this._projectionUtility.latLngToSvgPoint(d)[0])
-        .attr('cy', d => this._projectionUtility.latLngToSvgPoint(d)[1])
-        .on('mouseover', function(d, i){ self._onOverCircle(d, i, this);})
-        .on('mouseout', function(d){ self._onOutCircle(d, this);})
-        .on('touchstart', function(d, i){ self._onOverCircle(d, i, this);})
-        .on('touchleave', function(d){ self._onOutCircle(d, this);})
-        .on('mousedown', () => this._hideTooltip())
-        .call(this._onDragCircle());
+      .style('filter', 'url(#circle-shadow)')
+      .attr('class', getClass('cover-circle', this._options.invertColor))
+      .attr('r', 5)
+      .attr('cx', ([d]) => this._projectionUtility.latLngToSvgPoint(d)[0])
+      .attr('cy', ([d]) => this._projectionUtility.latLngToSvgPoint(d)[1])
+      .on('mouseover', function (event, [d, i]) {
+        self._onOverCircle(d, i, this);
+      })
+      .on('mouseout', function (event, [d]) {
+        self._onOutCircle(d, this);
+      })
+      .on('touchstart', function (event, [d, i]) {
+        event.preventDefault();
+        self._onOverCircle(d, i, this, true);
+      })
+      .on('touchend', function (event, [d]) {
+        event.preventDefault();
+        self._onOutCircle(d, this);
+      })
+      .on('mousedown', () => this._hideTooltip())
+      .call(this._onDragCircle());
 
-    this._nodeCircles.selectAll(".removed-circle").remove();
+    this._nodeCircles.selectAll('.removed-circle').remove();
   }
 
   _updateLine() {
     this._segments = [];
 
     let linesBase = this._linesBase
-      .selectAll("line")
-      .data(this._geometry ? this._geometry.lines : [])
-        .attr("class", "base-line")
-        .attr('x1', d => this._projectionUtility.latLngToSvgPoint(d[0])[0])
-        .attr('y1', d => this._projectionUtility.latLngToSvgPoint(d[0])[1])
-        .attr('x2', d => this._projectionUtility.latLngToSvgPoint(d[1])[0])
-        .attr('y2', d => this._projectionUtility.latLngToSvgPoint(d[1])[1])
-        .each(d => this._updateSegment(d));
+      .selectAll('line')
+      .data(this._geometry.lines)
+      .attr('class', getClass('base-line', this._options.invertColor))
+      .attr('x1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[0])
+      .attr('y1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[1])
+      .attr('x2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[0])
+      .attr('y2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[1])
+      .each((d) => this._updateSegment(d));
 
+    linesBase.exit().remove();
     linesBase
       .enter()
       .append('line')
-        .attr("class", "base-line")
-        .attr('x1', d => this._projectionUtility.latLngToSvgPoint(d[0])[0])
-        .attr('y1', d => this._projectionUtility.latLngToSvgPoint(d[0])[1])
-        .attr('x2', d => this._projectionUtility.latLngToSvgPoint(d[1])[0])
-        .attr('y2', d => this._projectionUtility.latLngToSvgPoint(d[1])[1])
-        .each(d => this._updateSegment(d));
-
-    linesBase.exit().remove();
+      .attr('class', getClass('base-line', this._options.invertColor))
+      .attr('x1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[0])
+      .attr('y1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[1])
+      .attr('x2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[0])
+      .attr('y2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[1])
+      .each((d) => this._updateSegment(d));
 
     let linesAux = this._linesAux
-      .selectAll("line")
-      .data(this._geometry ? this._geometry.lines : [])
-        .attr("class", "aux-line")
-        .attr('x1', d => this._projectionUtility.latLngToSvgPoint(d[0])[0])
-        .attr('y1', d => this._projectionUtility.latLngToSvgPoint(d[0])[1])
-        .attr('x2', d => this._projectionUtility.latLngToSvgPoint(d[1])[0])
-        .attr('y2', d => this._projectionUtility.latLngToSvgPoint(d[1])[1])
-      .on('mousemove', d => {
+      .selectAll('line')
+      .data(this._geometry.lines)
+      .join('line')
+      .datum((d, i) => [d, i])
+      .attr('class', 'aux-line')
+      .attr('x1', ([d]) => this._projectionUtility.latLngToSvgPoint(d[0])[0])
+      .attr('y1', ([d]) => this._projectionUtility.latLngToSvgPoint(d[0])[1])
+      .attr('x2', ([d]) => this._projectionUtility.latLngToSvgPoint(d[1])[0])
+      .attr('y2', ([d]) => this._projectionUtility.latLngToSvgPoint(d[1])[1]);
+
+    linesAux
+      .on('mousemove', (event, [d]) => {
         let point = Helper.findTouchPoint(
-          [this._projectionUtility.latLngToSvgPoint(d[0]), this._projectionUtility.latLngToSvgPoint(d[1])],
-          [event.offsetX,  event.offsetY]);
+          [
+            this._projectionUtility.latLngToSvgPoint(d[0]),
+            this._projectionUtility.latLngToSvgPoint(d[1]),
+          ],
+          [event.offsetX, event.offsetY]
+        );
         this._updateHoverCirclePosition(point[0], point[1]);
       })
-      .on('mouseout', d => this._hideHoverCircle())
+      .on('mouseout', () => this._hideHoverCircle())
       .on('mousedown', () => this._hideTooltip())
-      .call(this._onDragLine());
+      .call(this._onDragLine(linesAux, linesBase));
 
+    linesAux.exit().remove();
     linesAux
       .enter()
       .append('line')
-        .attr("class", "aux-line")
-        .attr('x1', d => this._projectionUtility.latLngToSvgPoint(d[0])[0])
-        .attr('y1', d => this._projectionUtility.latLngToSvgPoint(d[0])[1])
-        .attr('x2', d => this._projectionUtility.latLngToSvgPoint(d[1])[0])
-        .attr('y2', d => this._projectionUtility.latLngToSvgPoint(d[1])[1])
-        .on('mousemove', d => {
-          let point = Helper.findTouchPoint(
-            [this._projectionUtility.latLngToSvgPoint(d[0]), this._projectionUtility.latLngToSvgPoint(d[1])],
-            [event.offsetX,  event.offsetY]);
-          this._updateHoverCirclePosition(point[0], point[1]);
-        })
-        .on('mouseout', d => this._hideHoverCircle())
-        .on('mousedown', () => this._hideTooltip())
-        .call(this._onDragLine());
+      .join('line')
+      .datum((d, i) => [d, i])
+      .attr('class', 'aux-line')
+      .attr('x1', ([d]) => this._projectionUtility.latLngToSvgPoint(d[0])[0])
+      .attr('y1', ([d]) => this._projectionUtility.latLngToSvgPoint(d[0])[1])
+      .attr('x2', ([d]) => this._projectionUtility.latLngToSvgPoint(d[1])[0])
+      .attr('y2', ([d]) => this._projectionUtility.latLngToSvgPoint(d[1])[1]);
 
-    linesAux.exit().remove();
+    const lineDrag = this._lineDrag.selectAll('line').data([]);
+
+    lineDrag.exit().remove();
   }
 
   _updateSegmentText() {
-    let text = this._segmentText.selectAll("text")
-      .data(this._geometry ? this._geometry.lines : [])
-      .attr('class', 'segment-measure-text')
+    let text = this._segmentText
+      .selectAll('text')
+      .data(this._geometry.lines)
+      .attr('class', getClass('segment-measure-text', this._options.invertColor))
       .attr('text-anchor', 'middle')
       .attr('dominant-baseline', 'text-before-edge')
-      .attr('transform', d => {
+      .attr('transform', (d) => {
         let p1 = this._projectionUtility.latLngToSvgPoint(d[0]);
         let p2 = this._projectionUtility.latLngToSvgPoint(d[1]);
         return Helper.transformText(p1, p2);
       })
-      .text((d, i) => this._helper.formatLength(this._helper.computeLengthBetween(d[0], d[1])));
+      .text((d, i) =>
+        this._helper.formatLength(this._helper.computeLengthBetween(d[0], d[1]))
+      );
 
-    text.enter()
+    text
+      .enter()
       .append('text')
-      .attr('class', 'segment-measure-text')
+      .attr('class', getClass('segment-measure-text', this._options.invertColor))
       .attr('text-anchor', 'middle')
       .attr('dominant-baseline', 'text-before-edge')
-      .attr('transform', d => {
+      .attr('transform', (d) => {
         let p1 = this._projectionUtility.latLngToSvgPoint(d[0]);
         let p2 = this._projectionUtility.latLngToSvgPoint(d[1]);
         return Helper.transformText(p1, p2);
       })
-      .text((d, i) => this._helper.formatLength(this._helper.computeLengthBetween(d[0], d[1])));
+      .text((d, i) =>
+        this._helper.formatLength(this._helper.computeLengthBetween(d[0], d[1]))
+      );
 
     text.exit().remove();
   }
 
   _updateNodeText() {
-    let text = this._nodeText.selectAll("text")
-      .data(this._geometry ? this._geometry.nodes : [])
-      .attr('class', (d, i) => i === 0 ? 'node-measure-text head-text' : 'node-measure-text')
+    let text = this._nodeText
+      .selectAll('text')
+      .data(this._geometry.nodes)
+      .attr('class', (d, i) =>
+        i === 0 ? `${getClass('node-measure-text', this._options.invertColor)} head-text` : getClass('node-measure-text', this._options.invertColor)
+      )
       .attr('text-anchor', 'middle')
       .attr('dominant-baseline', 'text-after-edge')
-      .attr('x', d => this._projectionUtility.latLngToSvgPoint(d)[0])
+      .attr('x', (d) => this._projectionUtility.latLngToSvgPoint(d)[0])
       .attr('y', this._transformNodeTextY.bind(this))
       .text((d, i) => {
-        let len = this._helper.computePathLength(this._geometry.nodes.slice(0, i + 1));
+        let len = this._helper.computePathLength(
+          this._geometry.nodes.slice(0, i + 1)
+        );
         if (i === this._geometry.nodes.length - 1) {
           this._length = len;
         }
         return this._helper.formatLength(len);
       });
 
-    text.enter()
+    text
+      .enter()
       .append('text')
-      .attr('class', (d, i) => i === 0 ? 'node-measure-text head-text' : 'node-measure-text')
+      .attr('class', (d, i) =>
+        i === 0 ? `${getClass('node-measure-text', this._options.invertColor)} head-text` : getClass('node-measure-text', this._options.invertColor)
+      )
       .attr('text-anchor', 'middle')
       .attr('dominant-baseline', 'text-after-edge')
-      .attr('x', d => this._projectionUtility.latLngToSvgPoint(d)[0])
+      .attr('x', (d) => this._projectionUtility.latLngToSvgPoint(d)[0])
       .attr('y', this._transformNodeTextY.bind(this))
       .text((d, i) => {
-        let len = this._helper.computePathLength(this._geometry.nodes.slice(0, i + 1));
+        let len = this._helper.computePathLength(
+          this._geometry.nodes.slice(0, i + 1)
+        );
         if (i === this._geometry.nodes.length - 1) {
           this._length = len;
         }
@@ -477,12 +587,16 @@ export default class MeasureTool {
     text.exit().remove();
   }
 
-  _onOverCircle(d, i, target) {
+  _onOverCircle(d, i, target, isTouch = false) {
     if (this._dragging) return;
     select(target).attr('r', 6);
-    if (this._options.tooltip) {
-      this._tooltip.show(this._projectionUtility.latLngToContainerPoint(d),
-        i === 0 ? Config.tooltipText2 : Config.tooltipText1);
+    if (this._options.tooltip && !isTouch) {
+      this._tooltip.show(
+        this._projectionUtility.latLngToContainerPoint(d),
+        i === 0
+          ? Config.tooltipText2(this._options.language)
+          : Config.tooltipText1(this._options.language)
+      );
     }
   }
 
@@ -495,32 +609,32 @@ export default class MeasureTool {
     let self = this;
     let isDragged = false;
 
-    let circleDrag = drag()
-      .on('drag', function (d, i) {
-        isDragged = true;
-        self._dragging = true;
+    let circleDrag = drag().on('drag', function (event, [, i]) {
+      isDragged = true;
+      self._dragging = true;
 
-        select(this)
-          .attr('cx', event.x)
-          .attr('cy', event.y);
-        self._updateLinePosition.call(self, self._linesBase, i);
-        self._updateLinePosition.call(self, self._linesAux, i);
-        if (self._options.showSegmentLength) {
-          self._updateSegmentTextPosition(i);
-        }
-        if (self._options.showAccumulativeLength) {
-          self._updateNodeTextPosition(i);
-        }
-        self._updateArea(i, self._projectionUtility.svgPointToLatLng([event.x, event.y]));
-      });
+      select(this).attr('cx', event.x).attr('cy', event.y);
+      self._updateLinePosition.call(self, self._linesBase, i, event);
+      self._updateLinePosition.call(self, self._linesAux, i, event);
+      if (self._options.showSegmentLength) {
+        self._updateSegmentTextPosition(i, event);
+      }
+      if (self._options.showAccumulativeLength) {
+        self._updateNodeTextPosition(i, event);
+      }
+      self._updateArea(
+        i,
+        self._projectionUtility.svgPointToLatLng([event.x, event.y])
+      );
+    });
 
-    circleDrag.on('start', function(d) {
+    circleDrag.on('start', function (event) {
       event.sourceEvent.stopPropagation();
       select(this).raise().attr('r', 6);
       self._disableMapScroll();
     });
 
-    circleDrag.on('end', function (d, i) {
+    circleDrag.on('end', function (event, [d, i]) {
       self._enableMapScroll();
       if (!isDragged) {
         if (i > 0) {
@@ -533,8 +647,14 @@ export default class MeasureTool {
       } else {
         self._geometry.updateNode(
           i,
-          self._projectionUtility.svgPointToLatLng([event.x, event.y]));
-        self._showTooltipOnEvent(i === 0 ? Config.tooltipText2 : Config.tooltipText1);
+          self._projectionUtility.svgPointToLatLng([event.x, event.y])
+        );
+        self._showTooltipOnEvent(
+          i === 0
+            ? Config.tooltipText2(self._options.language)
+            : Config.tooltipText1(self._options.language),
+          event
+        );
       }
       isDragged = false;
       self._dragging = false;
@@ -546,165 +666,228 @@ export default class MeasureTool {
 
   _onDragLine() {
     let isDragged = false;
-    let lineDrag = drag()
-      .on('drag', (d, i) => {
-        this._dragging = true;
-        if (!isDragged) {
-          isDragged = true;
-          this._geometry.insertNode(
-            i + 1,
-            this._projectionUtility.svgPointToLatLng([event.x, event.y]));
-          this._updateLine();
-          if (this._options.showSegmentLength) {
-            this._updateSegmentText();
-          }
-          if (this._options.showAccumulativeLength) {
-            this._updateNodeText();
-          }
-        }
-        this._updateHoverCirclePosition(event.x, event.y);
-        this._updateLinePosition(this._linesBase, i + 1);
-        this._updateLinePosition(this._linesAux, i + 1);
+    let lineDrag = drag().on('drag', (event, [, i]) => {
+      this._dragging = true;
+      if (!isDragged) {
+        isDragged = true;
+
+        // idea - The problems seems that we can't rejoin the data on the same selection during the drag,
+        // but we can use the new data to create a new line just to show during the drag by passing the drag event to it.
+        this._geometry.insertNode(
+          i + 1,
+          this._projectionUtility.svgPointToLatLng([event.x, event.y])
+        );
+
+        const lineDrag = this._lineDrag
+          .selectAll('line')
+          .data(this._geometry.lines);
+
+        lineDrag.exit().remove();
+        lineDrag
+          .enter()
+          .append('line')
+          .attr('class', getClass('base-line', this._options.invertColor))
+          .attr('x1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[0])
+          .attr('y1', (d) => this._projectionUtility.latLngToSvgPoint(d[0])[1])
+          .attr('x2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[0])
+          .attr('y2', (d) => this._projectionUtility.latLngToSvgPoint(d[1])[1]);
+
+        this._linesBase.selectAll('line').style('display', 'none');
+        this._linesAux.selectAll('line').style('display', 'none');
+
         if (this._options.showSegmentLength) {
-          this._updateSegmentTextPosition(i + 1);
+          this._updateSegmentText();
         }
         if (this._options.showAccumulativeLength) {
-          this._updateNodeTextPosition(i + 1);
+          this._updateNodeText();
         }
-        this._updateArea(i + 1, this._projectionUtility.svgPointToLatLng([event.x, event.y]));
-      });
+      }
+      this._updateHoverCirclePosition(event.x, event.y);
+      this._updateLinePosition(this._lineDrag, i + 1, event);
+      if (this._options.showSegmentLength) {
+        this._updateSegmentTextPosition(i + 1, event);
+      }
+      if (this._options.showAccumulativeLength) {
+        this._updateNodeTextPosition(i + 1, event);
+      }
+      this._updateArea(
+        i + 1,
+        this._projectionUtility.svgPointToLatLng([event.x, event.y])
+      );
+    });
 
-    lineDrag.on('start', () => {
+    lineDrag.on('start', (event) => {
       event.sourceEvent.stopPropagation();
-      this._hoverCircle.select("circle")
-        .attr('class', "cover-circle");
+      this._hoverCircle.select('circle').attr('class', getClass('cover-circle', this._options.invertColor));
       this._disableMapScroll();
     });
 
-    lineDrag.on('end', (d, i) => {
+    lineDrag.on('end', (event, [, i]) => {
       this._enableMapScroll();
       if (isDragged) {
         this._geometry.updateNode(
           i + 1,
-          this._projectionUtility.svgPointToLatLng([event.x, event.y]));
+          this._projectionUtility.svgPointToLatLng([event.x, event.y])
+        );
         this._hideHoverCircle();
         this._overlay.draw();
         isDragged = false;
-        this._showTooltipOnEvent(Config.tooltipText1);
+        this._showTooltipOnEvent(
+          Config.tooltipText1(this._options.language),
+          event
+        );
       }
-      this._updateArea(i + 1, this._projectionUtility.svgPointToLatLng([event.x, event.y]));
-      this._hoverCircle.select("circle")
-        .attr('class', "grey-circle");
+      this._updateArea(
+        i + 1,
+        this._projectionUtility.svgPointToLatLng([event.x, event.y])
+      );
+      this._hoverCircle.select('circle')
+        .style('filter', 'url(#circle-shadow)')
+        .attr('class', getClass('grey-circle', this._options.invertColor));
+      this._linesBase.selectAll('line').style('display', 'inline');
+      this._linesAux.selectAll('line').style('display', 'inline');
       this._dragging = false;
     });
 
     return lineDrag;
   }
 
-  _updateLinePosition(group, i) {
+  _updateLinePosition(group, i, event) {
     if (i < this._geometry.lines.length) {
-      group.select(`line:nth-child(${i + 1})`)
+      group
+        .select(`line:nth-child(${i + 1})`)
         .attr('x1', event.x)
         .attr('y1', event.y);
     }
     if (i > 0) {
-      group.select(`line:nth-child(${i})`)
+      group
+        .select(`line:nth-child(${i})`)
         .attr('x2', event.x)
         .attr('y2', event.y);
     }
   }
 
-  _updateSegmentTextPosition(index) {
+  _updateSegmentTextPosition(index, event) {
     if (index < this._geometry.lines.length) {
-      this._segmentText.select(`text:nth-child(${index + 1})`)
-        .attr('transform', d => {
+      this._segmentText
+        .select(`text:nth-child(${index + 1})`)
+        .attr('transform', (d) => {
           let p1 = [event.x, event.y];
           let p2 = this._projectionUtility.latLngToSvgPoint(d[1]);
           return Helper.transformText(p1, p2);
         })
-        .text(d => this._helper.formatLength(this._helper.computeLengthBetween(
-          this._projectionUtility.svgPointToLatLng([event.x, event.y]), d[1])));
+        .text((d) =>
+          this._helper.formatLength(
+            this._helper.computeLengthBetween(
+              this._projectionUtility.svgPointToLatLng([event.x, event.y]),
+              d[1]
+            )
+          )
+        );
     }
     if (index > 0) {
-      this._segmentText.select(`text:nth-child(${index})`)
-        .attr('transform', d => {
+      this._segmentText
+        .select(`text:nth-child(${index})`)
+        .attr('transform', (d) => {
           let p1 = this._projectionUtility.latLngToSvgPoint(d[0]);
           let p2 = [event.x, event.y];
           return Helper.transformText(p1, p2);
         })
-        .text(d => this._helper.formatLength(this._helper.computeLengthBetween(
-          d[0], this._projectionUtility.svgPointToLatLng([event.x, event.y]))));
+        .text((d) =>
+          this._helper.formatLength(
+            this._helper.computeLengthBetween(
+              d[0],
+              this._projectionUtility.svgPointToLatLng([event.x, event.y])
+            )
+          )
+        );
     }
   }
 
-  _updateNodeTextPosition(index) {
-    this._nodeText.select(`text:nth-child(${index + 1})`)
+  _updateNodeTextPosition(index, event) {
+    this._nodeText
+      .select(`text:nth-child(${index + 1})`)
       .attr('x', event.x)
       .attr('y', () => {
         let offset;
-        if (index > 0 && this._projectionUtility.latLngToSvgPoint(
-            this._geometry.nodes[index - 1])[1] < event.y) {
+        if (
+          index > 0 &&
+          this._projectionUtility.latLngToSvgPoint(
+            this._geometry.nodes[index - 1]
+          )[1] < event.y
+        ) {
           offset = 23;
         } else {
           offset = -7;
         }
         return event.y + offset;
       });
-    this._nodeText.select(`text:nth-child(${index + 2})`)
-      .attr('y', d => {
-        let offset;
-        if (index + 1 > 0 && event.y < this._projectionUtility.latLngToSvgPoint(d)[1]) {
-          offset = 23;
-        } else {
-          offset = -7;
-        }
-        return this._projectionUtility.latLngToSvgPoint(d)[1] + offset;
-      });
-    let followingNodes = this._nodeText.selectAll('text')
+    this._nodeText.select(`text:nth-child(${index + 2})`).attr('y', (d) => {
+      let offset;
+      if (
+        index + 1 > 0 &&
+        event.y < this._projectionUtility.latLngToSvgPoint(d)[1]
+      ) {
+        offset = 23;
+      } else {
+        offset = -7;
+      }
+      return this._projectionUtility.latLngToSvgPoint(d)[1] + offset;
+    });
+    let followingNodes = this._nodeText
+      .selectAll('text')
       .filter((d, i) => i >= index);
-    followingNodes
-      .text((d, i) => {
-        let len = this._helper.computePathLength([
-          ...this._geometry.nodes.slice(0, index),
-          this._projectionUtility.svgPointToLatLng([event.x, event.y]),
-          ...this._geometry.nodes.slice(index + 1, index + 1 + i)
-        ]);
-        if (index + i === this._geometry.nodes.length - 1) {
-          this._length = len;
-        }
-        return this._helper.formatLength(len);
-      });
+    followingNodes.text((d, i) => {
+      let len = this._helper.computePathLength([
+        ...this._geometry.nodes.slice(0, index),
+        this._projectionUtility.svgPointToLatLng([event.x, event.y]),
+        ...this._geometry.nodes.slice(index + 1, index + 1 + i),
+      ]);
+      if (index + i === this._geometry.nodes.length - 1) {
+        this._length = len;
+      }
+      return this._helper.formatLength(len);
+    });
   }
 
   _updateHoverCirclePosition(x, y) {
-    this._hoverCircle.select("circle")
-      .attr('cx', x)
-      .attr('cy', y);
+    this._hoverCircle.select('circle').attr('cx', x).attr('cy', y);
     if (this._dragging) return;
     if (this._options.tooltip) {
-      this._tooltip.show(this._projectionUtility.svgPointToContainerPoint([x, y]), Config.tooltipText2);
+      this._tooltip.show(
+        this._projectionUtility.svgPointToContainerPoint([x, y]),
+        Config.tooltipText2(this._options.language)
+      );
     }
   }
 
   _hideHoverCircle() {
-    this._hoverCircle.select("circle")
-      .attr('cx', null)
-      .attr('cy', null);
+    this._hoverCircle.select('circle').attr('cx', null).attr('cy', null);
     this._hideTooltip();
   }
 
   _disableMapScroll() {
-    this._zoomControl = !!document.querySelector("button[aria-label='Zoom in']");
-    this._map.setOptions({ scrollwheel: false, gestureHandling: "none", zoomControl: false });
+    this._zoomControl = !!document.querySelector(
+      "button[aria-label='Zoom in']"
+    );
+    this._map.setOptions({
+      scrollwheel: false,
+      gestureHandling: 'none',
+      zoomControl: false,
+    });
   }
 
   _enableMapScroll() {
-    this._map.setOptions({ scrollwheel: true, gestureHandling: "auto", zoomControl: this._zoomControl });
+    this._map.setOptions({
+      scrollwheel: true,
+      gestureHandling: 'auto',
+      zoomControl: this._zoomControl,
+    });
   }
 
   _transformNodeTextY(d, i) {
     let offset;
-      if (i > 0 && this._geometry.nodes[i - 1][1] > d[1]) {
+    if (i > 0 && this._geometry.nodes[i - 1][1] > d[1]) {
       offset = 23;
     } else {
       offset = -7;
@@ -715,48 +898,75 @@ export default class MeasureTool {
   _updateArea(i, pointToCompare) {
     if (!this._geometry) return;
     const n = this._geometry.nodes.length;
-    const tolerance = 1 / 80 *  this.length;
-    let offset, area = 0;
+    const tolerance = (1 / 80) * this.length;
+    let offset,
+      area = 0;
     if (n > 2) {
       if (i === 0) {
         offset = this._helper.computeLengthBetween(
           this._geometry.nodes[n - 1],
-          pointToCompare);
-        area = offset > tolerance ? 0 : this._helper.computeArea([
-            pointToCompare,
-            ...this._geometry.nodes.slice(1, n - 1)
-          ]);
+          pointToCompare
+        );
+        area =
+          offset > tolerance
+            ? 0
+            : this._helper.computeArea([
+              pointToCompare,
+              ...this._geometry.nodes.slice(1, n - 1),
+            ]);
       } else if (i === n - 1) {
         offset = this._helper.computeLengthBetween(
           pointToCompare,
-          this._geometry.nodes[0]);
-        area = offset > tolerance ? 0 : this._helper.computeArea(this._geometry.nodes.slice(0, n - 1));
+          this._geometry.nodes[0]
+        );
+        area =
+          offset > tolerance
+            ? 0
+            : this._helper.computeArea(this._geometry.nodes.slice(0, n - 1));
       } else if (i > 0 && i < n - 1) {
         offset = this._helper.computeLengthBetween(
           this._geometry.nodes[0],
-          this._geometry.nodes[n - 1]);
-        area = offset > tolerance ? 0 : this._helper.computeArea([
-            ...this._geometry.nodes.slice(0, i),
-            pointToCompare,
-            ...this._geometry.nodes.slice(i + 1)
-          ]);
+          this._geometry.nodes[n - 1]
+        );
+        area =
+          offset > tolerance
+            ? 0
+            : this._helper.computeArea([
+              ...this._geometry.nodes.slice(0, i),
+              pointToCompare,
+              ...this._geometry.nodes.slice(i + 1),
+            ]);
       } else {
         offset = this._helper.computeLengthBetween(
           this._geometry.nodes[0],
-          this._geometry.nodes[n - 1]);
-        area = offset > tolerance ? 0 : this._helper.computeArea(this._geometry.nodes);
+          this._geometry.nodes[n - 1]
+        );
+        area =
+          offset > tolerance
+            ? 0
+            : this._helper.computeArea(this._geometry.nodes);
       }
     }
     this._area = area;
     if (area > 0) {
-      this._nodeText.select(':last-child')
-        .text(`Total distance: ${this.lengthText}; Total area: ${this.areaText}.`);
+      this._nodeText
+        .select(':last-child')
+        .text(
+          `Total distance: ${this.lengthText}; Total area: ${this.areaText}.`
+        );
     }
   }
 
-  _showTooltipOnEvent(text) {
+  _showTooltipOnEvent(text, event) {
+    // don't show tooltip in a touch event
+    if (event.sourceEvent.type.startsWith('touch')) {
+      return;
+    }
     if (this._options.tooltip) {
-      this._tooltip.show(this._projectionUtility.svgPointToContainerPoint([event.x, event.y]), text);
+      this._tooltip.show(
+        this._projectionUtility.svgPointToContainerPoint([event.x, event.y]),
+        text
+      );
     }
   }
 
@@ -770,28 +980,33 @@ export default class MeasureTool {
     if (!this._started) return;
     const result = this._getResults();
 
-    if (this._lastMeasure && this._lastMeasure.result.lengthText === this.lengthText && this._lastMeasure.result.areaText === this.areaText) return;
-    if (typeof this._events.get(EVENT_CHANGE) === "function") {
-      this._events.get(EVENT_CHANGE)(this._lastMeasure = result);
+    if (
+      this._lastMeasure &&
+      this._lastMeasure.result.lengthText === this.lengthText &&
+      this._lastMeasure.result.areaText === this.areaText
+    )
+      return;
+    if (typeof this._events.get(EVENT_CHANGE) === 'function') {
+      this._events.get(EVENT_CHANGE)((this._lastMeasure = result));
     }
   }
 
   _updateSegment(d) {
     const len = this._helper.computeLengthBetween(d[0], d[1]);
     const lenTxt = this._helper.formatLength(len);
-    this._segments.push(new Segment(d[0], d[1], len,lenTxt).toJSON());
+    this._segments.push(new Segment(d[0], d[1], len, lenTxt).toJSON());
   }
 
   _getResults() {
-    return ({
+    return {
       result: {
         length: this.length,
         lengthText: this.lengthText,
         area: this.area,
         areaText: this.areaText,
         segments: this.segments,
-        points: this.points
-      }
-    });
+        points: this.points,
+      },
+    };
   }
-};
+}