+++ src/polyfills.js
@@ -1,6 +1,6 @@
-export const ObjectAssign = function() {
+export const ObjectAssign = function () {
   if (typeof Object.assign !== 'function') {
-    Object.assign = function(target) {
+    Object.assign = function (target) {
       'use strict';
       if (target === null) {
         throw new TypeError('Cannot convert undefined or null to object');