+++ src/projection-utility.js
@@ -1,7 +1,7 @@
 export default class ProjectionUtility {
   constructor(div, projection, options) {
     this._defaultOptions = {
-      offsetRate: 8000
+      offsetRate: 8000,
     };
     this._options = Object.assign({}, this._defaultOptions, options || {});
     this._container = div;
@@ -16,7 +16,7 @@ export default class ProjectionUtility {
   }
 
   svgPointToLatLng(point) {
-    let rate = this._options.offsetRate / 2 ;
+    let rate = this._options.offsetRate / 2;
     let svgPoint = new google.maps.Point(point[0] - rate, point[1] - rate);
     let coords = this._projection.fromDivPixelToLatLng(svgPoint);
     return [coords.lng(), coords.lat()];
@@ -25,11 +25,13 @@ export default class ProjectionUtility {
   svgPointToContainerPoint(point) {
     let svgPoint = this.svgPointToLatLng(point);
     return this._projection.fromLatLngToContainerPixel(
-      new google.maps.LatLng(svgPoint[1], svgPoint[0]));
+      new google.maps.LatLng(svgPoint[1], svgPoint[0])
+    );
   }
 
   latLngToContainerPoint(coords) {
     return this._projection.fromLatLngToContainerPixel(
-      new google.maps.LatLng(coords[1], coords[0]));
+      new google.maps.LatLng(coords[1], coords[0])
+    );
   }
-};
+}